Colocate generic List event args in non-generic files
